added deleted universal products

+++ src/modules/admin_actions/handlers/editor/category/delete_handlers.py
@@ -10,6 +10,9 @@ from src.modules.admin_actions.keyboards import delete_category_kb, back_in_cate
 from src.modules.admin_actions.services import safe_get_category, upload_category
 from src.services.database.categories.actions import delete_category as delete_category_service, \
     delete_product_accounts_by_category
+from src.services.database.categories.actions.products.universal.action_delete import \
+    delete_product_universal_by_category
+from src.services.database.categories.models.main_category_and_product import ProductType
 from src.services.database.users.models import Users
 from src.utils.i18n import get_text
 
@@ -86,7 +89,10 @@ async def delete_all_products(callback: CallbackQuery, user: Users):
         return
 
     await upload_category(category, user, callback)
-    await delete_product_accounts_by_category(category_id)
+    if category.product_type == ProductType.ACCOUNT:
+        await delete_product_accounts_by_category(category_id)
+    if category.product_type == ProductType.UNIVERSAL:
+        await delete_product_universal_by_category(category_id)
 
     await callback.answer("Products successfully deleted", show_alert=True)
     await show_category(user, category_id, send_new_message=True)
